requirements.txt
flask
torch
torchvision
opencv-python
numpy
scipy
Pillow
matplotlib
seaborn
pandas
